debundle/lowering: extract imports_cross
BAZEL_TEST_INVOCATIONS=none: file move.
